MOR-452: Moved the desktop bar editor into a component.
