+++ src/styles/app.scss
@@ -64,5 +64,9 @@ body.isMobile .main {
   }
 }
 
+.max-height {
+  height: 100%;
+}
+
 @import "./expander";
